requirements.txt
flask
flask_sqlalchemy
python-dotenv
pymysql
redis
flask-jwt-extended
celery
requests
pytest
jieba
click
flask-migrate